Move resolve_imports_and_packages_back
