+++ src/snowflake/snowpark/_internal/udf_utils.py
@@ -1,15 +1,18 @@
 #
 # Copyright (c) 2012-2023 Snowflake Computing Inc. All rights reserved.
 #
-
+import io
 import os
 import pickle
 import sys
 import typing
+import zipfile
 from logging import getLogger
+from types import ModuleType
 from typing import (
     Any,
     Callable,
+    Dict,
     List,
     NamedTuple,
     Optional,
@@ -33,7 +36,12 @@ from snowflake.snowpark._internal.type_utils import (
 from snowflake.snowpark._internal.utils import (
     STAGE_PREFIX,
     TempObjectType,
+    get_udf_upload_prefix,
+    is_single_quoted,
+    normalize_remote_file_or_dir,
     random_name_for_temp_object,
+    random_number,
+    unwrap_stage_location_single_quote,
     validate_object_name,
 )
 from snowflake.snowpark.types import (
@@ -508,6 +516,153 @@ def {_DEFAULT_HANDLER_NAME}({args}):
 {deserialization_code}
 {func_code}
 """.strip()
+
+
+def resolve_imports_and_packages(
+    session: "snowflake.snowpark.Session",
+    object_type: TempObjectType,
+    func: Union[Callable, Tuple[str, str]],
+    arg_names: List[str],
+    udf_name: str,
+    stage_location: Optional[str],
+    imports: Optional[List[Union[str, Tuple[str, str]]]],
+    packages: Optional[List[Union[str, ModuleType]]],
+    parallel: int = 4,
+    is_pandas_udf: bool = False,
+    is_dataframe_input: bool = False,
+    force_push: bool = True,
+    max_batch_size: Optional[int] = None,
+    *,
+    statement_params: Optional[Dict[str, str]] = None,
+    source_code_display: bool = False,
+    skip_upload_on_content_match: bool = False,
+) -> Tuple[str, str, str, str, str]:
+    upload_stage = (
+        unwrap_stage_location_single_quote(stage_location)
+        if stage_location
+        else session.get_session_stage()
+    )
+
+    # resolve imports
+    if imports:
+        udf_level_imports = {}
+        for udf_import in imports:
+            if isinstance(udf_import, str):
+                resolved_import_tuple = session._resolve_import_path(udf_import)
+            elif isinstance(udf_import, tuple) and len(udf_import) == 2:
+                resolved_import_tuple = session._resolve_import_path(
+                    udf_import[0], udf_import[1]
+                )
+            else:
+                raise TypeError(
+                    f"{get_error_message_abbr(object_type).replace(' ', '-')}-level import can only be a file path (str) "
+                    "or a tuple of the file path (str) and the import path (str)."
+                )
+            udf_level_imports[resolved_import_tuple[0]] = resolved_import_tuple[1:]
+        all_urls = session._resolve_imports(
+            upload_stage, udf_level_imports, statement_params=statement_params
+        )
+    elif imports is None:
+        all_urls = session._resolve_imports(
+            upload_stage, statement_params=statement_params
+        )
+    else:
+        all_urls = []
+
+    # resolve packages
+    resolved_packages = (
+        session._resolve_packages(
+            packages, include_pandas=is_pandas_udf, force_push=force_push
+        )
+        if packages is not None
+        else session._resolve_packages(
+            [],
+            session._packages,
+            validate_package=False,
+            include_pandas=is_pandas_udf,
+            force_push=force_push,
+        )
+    )
+
+    dest_prefix = get_udf_upload_prefix(udf_name)
+
+    # Upload closure to stage if it is beyond inline closure size limit
+    if isinstance(func, Callable):
+        # generate a random name for udf py file
+        # and we compress it first then upload it
+        udf_file_name_base = f"udf_py_{random_number()}"
+        udf_file_name = f"{udf_file_name_base}.zip"
+        code = generate_python_code(
+            func,
+            arg_names,
+            object_type,
+            is_pandas_udf,
+            is_dataframe_input,
+            max_batch_size,
+            source_code_display=source_code_display,
+        )
+        if len(code) > _MAX_INLINE_CLOSURE_SIZE_BYTES:
+            dest_prefix = get_udf_upload_prefix(udf_name)
+            upload_file_stage_location = normalize_remote_file_or_dir(
+                f"{upload_stage}/{dest_prefix}/{udf_file_name}"
+            )
+            udf_file_name_base = os.path.splitext(udf_file_name)[0]
+            with io.BytesIO() as input_stream:
+                with zipfile.ZipFile(
+                    input_stream, mode="w", compression=zipfile.ZIP_DEFLATED
+                ) as zf:
+                    zf.writestr(f"{udf_file_name_base}.py", code)
+                session._conn.upload_stream(
+                    input_stream=input_stream,
+                    stage_location=upload_stage,
+                    dest_filename=udf_file_name,
+                    dest_prefix=dest_prefix,
+                    parallel=parallel,
+                    source_compression="DEFLATE",
+                    compress_data=False,
+                    overwrite=True,
+                    is_in_udf=True,
+                    skip_upload_on_content_match=skip_upload_on_content_match,
+                )
+            all_urls.append(upload_file_stage_location)
+            inline_code = None
+            handler = f"{udf_file_name_base}.{_DEFAULT_HANDLER_NAME}"
+        else:
+            inline_code = code
+            upload_file_stage_location = None
+            handler = _DEFAULT_HANDLER_NAME
+    else:
+        udf_file_name = os.path.basename(func[0])
+        # for a compressed file, it might have multiple extensions
+        # and we should remove all extensions
+        udf_file_name_base = udf_file_name.split(".")[0]
+        inline_code = None
+        handler = f"{udf_file_name_base}.{func[1]}"
+
+        if func[0].startswith(STAGE_PREFIX):
+            upload_file_stage_location = None
+            all_urls.append(func[0])
+        else:
+            upload_file_stage_location = normalize_remote_file_or_dir(
+                f"{upload_stage}/{dest_prefix}/{udf_file_name}"
+            )
+            session._conn.upload_file(
+                path=func[0],
+                stage_location=upload_stage,
+                dest_prefix=dest_prefix,
+                parallel=parallel,
+                compress_data=False,
+                overwrite=True,
+                skip_upload_on_content_match=skip_upload_on_content_match,
+            )
+            all_urls.append(upload_file_stage_location)
+
+    # build imports and packages string
+    all_imports = ",".join(
+        [url if is_single_quoted(url) else f"'{url}'" for url in all_urls]
+    )
+    all_packages = ",".join([f"'{package}'" for package in resolved_packages])
+    return handler, inline_code, all_imports, all_packages, upload_file_stage_location
 
 
 def create_python_udf_or_sp(
