+++ src/snowflake/snowpark/stored_procedure.py
@@ -12,7 +12,6 @@ from typing import Any, Callable, Dict, List, Optional, Tuple, Union
 import snowflake.snowpark
 from snowflake.connector import ProgrammingError
 from snowflake.snowpark._internal.error_message import SnowparkClientExceptionMessages
-from snowflake.snowpark._internal.packaging_utils import resolve_imports_and_packages
 from snowflake.snowpark._internal.telemetry import TelemetryField
 from snowflake.snowpark._internal.type_utils import convert_sp_to_sf_type
 from snowflake.snowpark._internal.udf_utils import (
@@ -25,6 +24,7 @@ from snowflake.snowpark._internal.udf_utils import (
     generate_call_python_sp_sql,
     process_file_path,
     process_registration_inputs,
+    resolve_imports_and_packages,
 )
 from snowflake.snowpark._internal.utils import TempObjectType
 from snowflake.snowpark.types import DataType, StructType
